Break out REST types to core
So that the types can be reused elsewhere (from Karaf commands)
