+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/repair/types/VirtualNodeState.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ericsson.bss.cassandra.ecchronos.rest.types;
+package com.ericsson.bss.cassandra.ecchronos.core.repair.types;
 
 import com.datastax.driver.core.Host;
 import com.ericsson.bss.cassandra.ecchronos.core.repair.state.VnodeRepairState;
