+++ core/src/test/java/com/ericsson/bss/cassandra/ecchronos/core/repair/types/TestCompleteRepairJob.java
@@ -12,11 +12,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ericsson.bss.cassandra.ecchronos.rest.types;
+package com.ericsson.bss.cassandra.ecchronos.core.repair.types;
 
 import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairJobView;
 import com.ericsson.bss.cassandra.ecchronos.core.repair.state.VnodeRepairState;
-import com.ericsson.bss.cassandra.ecchronos.rest.TestUtils;
+import com.ericsson.bss.cassandra.ecchronos.core.repair.TestUtils;
 import com.google.common.collect.ImmutableSet;
 import org.junit.Test;
