+++ pom.xml
@@ -95,6 +95,7 @@
         <org.apache.maven.plugins-maven-pmd-plugin.version>3.11.0</org.apache.maven.plugins-maven-pmd-plugin.version>
         <org.apache.maven.plugins-maven-resources-plugin.version>3.1.0</org.apache.maven.plugins-maven-resources-plugin.version>
         <org.apache.maven.plugins-maven-dependency-plugin.version>3.1.1</org.apache.maven.plugins-maven-dependency-plugin.version>
+        <org.apache.maven.plugins-maven-jar-plugin.version>2.6</org.apache.maven.plugins-maven-jar-plugin.version>
         <org.apache.felix.maven-bundle-plugin.version>3.2.0</org.apache.felix.maven-bundle-plugin.version>
         <com.mycila.license-maven-plugin.version>3.0</com.mycila.license-maven-plugin.version>
         <io.fabric8.docker-maven-plugin.version>0.26.0</io.fabric8.docker-maven-plugin.version>
@@ -562,6 +563,12 @@ limitations under the License.
                     <groupId>org.codehaus.mojo</groupId>
                     <artifactId>exec-maven-plugin</artifactId>
                     <version>${org.codehaus.mojo.exec-maven-plugin.version}</version>
+                </plugin>
+
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-jar-plugin</artifactId>
+                    <version>${org.apache.maven.plugins-maven-jar-plugin.version}</version>
                 </plugin>
             </plugins>
         </pluginManagement>
