+++ rest/src/test/java/com/ericsson/bss/cassandra/ecchronos/rest/TestRepairSchedulerRESTImpl.java
@@ -17,10 +17,11 @@ package com.ericsson.bss.cassandra.ecchronos.rest;
 import com.datastax.driver.core.Host;
 import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairJobView;
 import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairScheduler;
+import com.ericsson.bss.cassandra.ecchronos.core.repair.types.CompleteRepairJob;
+import com.ericsson.bss.cassandra.ecchronos.core.repair.types.ScheduledRepairJob;
+import com.ericsson.bss.cassandra.ecchronos.core.repair.types.VirtualNodeState;
+import com.ericsson.bss.cassandra.ecchronos.core.repair.TestUtils;
 import com.ericsson.bss.cassandra.ecchronos.core.utils.LongTokenRange;
-import com.ericsson.bss.cassandra.ecchronos.rest.types.CompleteRepairJob;
-import com.ericsson.bss.cassandra.ecchronos.rest.types.ScheduledRepairJob;
-import com.ericsson.bss.cassandra.ecchronos.rest.types.VirtualNodeState;
 import com.google.common.collect.ImmutableSet;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
